+++ src/main/php/net/stubbles/ioc/App.php
@@ -9,7 +9,6 @@
  */
 namespace net\stubbles\ioc;
 use net\stubbles\ioc\module\ModeBindingModule;
-use net\stubbles\ioc\module\PropertiesBindingModule;
 use net\stubbles\lang\Mode;
 /**
  * Application base class.
@@ -76,57 +75,6 @@ abstract class App
     }
 
     /**
-     * enable persistent annotation cache with given cache storage logic
-     *
-     * The $readCache closure must return the stored annotation data. If no such
-     * data is present it must return null. In case the stored annotation data
-     * can't be unserialized into an array a
-     * net\stubbles\lang\exception\RuntimeException will be thrown.
-     *
-     * The $storeCache closure must store passed annotation data. It doesn't
-     * need to take care about serialization, as it already receives a
-     * serialized representation.
-     *
-     * A possible implementation for the file cache would look like this:
-     * <code>
-     * self::persistAnnotations(function() use($cacheFile)
-     *                          {
-     *                              if (file_exists($cacheFile)) {
-     *                                  return file_get_contents($cacheFile);
-     *                              }
-     *
-     *                              return null;
-     *                          },
-     *                          function($annotationData) use($cacheFile)
-     *                          {
-     *                              file_put_contents($cacheFile, $annotationData);
-     *                          }
-     * );
-     * </code>
-     *
-     * @param  \Closure  $readCache
-     * @param  \Closure  $storeCache
-     * @since  3.0.0
-     * @deprecated  since 3.1.0, will be removed with 4.0.0, use \net\stubbles\lang\persistAnnotations() instead
-     */
-    protected static function persistAnnotations(\Closure $readCache, \Closure $storeCache)
-    {
-        \net\stubbles\lang\persistAnnotations($readCache, $storeCache);
-    }
-
-    /**
-     * enable persistent annotation cache by telling where to store cache data
-     *
-     * @param  string  $cacheFile
-     * @since  3.0.0
-     * @deprecated  since 3.1.0, will be removed with 4.0.0, use \net\stubbles\lang\persistAnnotationsInFile() instead
-     */
-    protected static function persistAnnotationsInFile($cacheFile)
-    {
-        \net\stubbles\lang\persistAnnotationsInFile($cacheFile);
-    }
-
-    /**
      * creates mode binding module
      *
      * @api
@@ -138,20 +86,6 @@ abstract class App
     protected static function createModeBindingModule($projectPath, Mode $mode = null)
     {
         return new ModeBindingModule($projectPath, $mode);
-    }
-
-    /**
-     * creates properties binding module
-     *
-     * @api
-     * @param   string  $projectPath
-     * @return  PropertiesBindingModule
-     * @since   2.0.0
-     * @deprecated  since 3.4.0, properties are now bound via createModeBindingModule()
-     */
-    protected static function createPropertiesBindingModule($projectPath)
-    {
-        return new PropertiesBindingModule($projectPath);
     }
 
     /**