+++ src/test/php/net/stubbles/ioc/AppTestCase.php
@@ -71,17 +71,6 @@ class AppTestCase extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @since  2.0.0
-     * @test
-     */
-    public function canCreatePropertiesBindingModule()
-    {
-        $this->assertInstanceOf('net\stubbles\ioc\module\PropertiesBindingModule',
-                                AppUsingBindingModule::getPropertiesBindingModule(__DIR__)
-        );
-    }
-
-    /**
      * @since  2.1.0
      * @group  issue_33
      * @test
@@ -92,63 +81,6 @@ class AppTestCase extends \PHPUnit_Framework_TestCase
                             AppClassWithBindings::create('projectPath')
                                                 ->wasBoundBy()
         );
-    }
-
-    /**
-     * creates a annotation cache with one annotation
-     *
-     * @return  string
-     * @deprecated  since 3.1.0, will be removed with 4.0.0
-     */
-    private function createdCachedAnnotation()
-    {
-        return serialize(array(Annotation::TARGET_CLASS => array('foo' => array('bar' => new Annotation('bar')))));
-    }
-
-    /**
-     * @deprecated  since 3.1.0, will be removed with 4.0.0
-     * @since  3.0.0
-     * @group  issue_58
-     * @test
-     */
-    public function canCreateAppInstanceWithFileAnnotationCache()
-    {
-        $root = vfsStream::setup();
-        $file = vfsStream::newFile('annotations.cache')
-                         ->withContent($this->createdCachedAnnotation())
-                         ->at($root);
-        AppUsingBindingModule::callAnnotationFilePersistence($file->url());
-        $this->assertTrue(AnnotationCache::has(Annotation::TARGET_CLASS, 'foo', 'bar'));
-
-    }
-
-    /**
-     * @deprecated  since 3.1.0, will be removed with 4.0.0
-     * @since  3.0.0
-     * @group  issue_58
-     * @test
-     */
-    public function canCreateAppInstanceWithOtherAnnotationCache()
-    {
-        $annotationData = $this->createdCachedAnnotation();
-        AppUsingBindingModule::callAnnotationPersistence(function() use($annotationData)
-                                                         {
-                                                             return $annotationData;
-
-                                                         },
-                                                         function($data) {}
-        );
-        $this->assertTrue(AnnotationCache::has(Annotation::TARGET_CLASS, 'foo', 'bar'));
-    }
-
-    /**
-     * clean up test environment
-     *
-     * @deprecated  since 3.1.0, will be removed with 4.0.0
-     */
-    public function tearDown()
-    {
-        AnnotationCache::stop();
     }
 
     /**